fix: PR changes addressed
