+++ src/app/signin/page.tsx
@@ -1,37 +1,8 @@
-'use client';
-
-import { useEffect, useState } from 'react';
-
-import { AppType } from '@/server/app';
-import { ProvidersType } from '@/features/auth/types/next-auth';
-import SignInForm from '@/features/auth/comps/SignInForm';
-import { getProviders } from 'next-auth/react';
-import { hc } from 'hono/client';
-
-type Props = {};
-
-const SigninPage = (props: Props) => {
-    const [providers, setProviders] = useState<ProvidersType | null>(null);
-    const [email, setEmail] = useState('');
-
-    useEffect(() => {
-        (async () => {
-            const oAuthProviders = await getProviders();
-            setProviders(oAuthProviders);
-        })();
-    }, []);
-
-    const handleSubmit = async () => {
-        const client = hc<AppType>(process.env.BASE_URL);
-        const res = await client.api.signin['magic-link'].$post({ json: { email } });
-        const data = await res.json();
-    };
-
+import SignInPage from '@/features/auth/comps/SignInPage';
+export default function SignIn() {
     return (
-        <main className="h-[calc(100vh_-_64px_-_10vh)] flex flex-col gap-12">
-            <SignInForm providers={providers} email={email} setEmail={setEmail} handleSubmit={handleSubmit} />
-        </main>
+        <>
+            <SignInPage />
+        </>
     );
-};
-
-export default SigninPage;
+}
