+++ src/features/auth/comps/SignInForm.tsx
@@ -2,8 +2,6 @@
 
 import OAuthProviderButton from '@/features/auth/comps/OAuthProviderButton';
 import { ProvidersType } from '../types/next-auth';
-import Image from 'next/image';
-import Github from '@/../public/icons/ic_github.svg';
 import { Button } from '@/components/ui/button';
 
 type Props = {
@@ -22,7 +20,7 @@ const SignInForm = ({ providers, email, setEmail, handleSubmit }: Props) => {
                 </span>
             </div>
             <div className="md:max-w-[1024px] flex flex-col mx-auto px-4 grow h-full w-full items-center justify-center">
-                <div className="glass-card-gradient md:shadow-[0px_4px_4px_rgba(0,0,0,0.04)] z-20 max-w-sm w-full min-h-[300px] rounded-xl md:border md:border-[#DEE5EE] p-6 flex items-center justify-center flex-col">
+                <div className="md:shadow-[0px_4px_4px_rgba(0,0,0,0.04)] z-20 max-w-sm w-full rounded-xl md:border md:border-[#DEE5EE] p-6 flex items-center justify-center flex-col">
                     <div className="flex flex-col font-satoshi text-center gap-2 font-medium text-base leading-[23px]">
                         <span> Welcome to</span>
                         <span className="font-clashGrotesk font-semibold text-2xl leading-[29px] text-[#0057CC]">
